+++ CHANGELOG.md
@@ -36,3 +36,4 @@
 ### 1.0.5 (2024-05-28)
 
 - Removed private key from fees estimation
+- Added hiro api key to estimate transaction fee